input: misc: cm36686: Update from 26.3.A.1.33
This is a sanitized catch-up for this driver, which fixes its
functionality and indentation/code style issues; and cleans up
the unused code, because its old version was suffering from
the mentioned problems.
I don't think this commit/code is messy and instead, it is
actually fair, because the driver looks quite better now, what
makes me think that FIH did not put its hands in here.
Anyway, I will be trying to improve the diff of this commit if
possible.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>