arm: DT: msm8939-mtp: Set an specific place for bma2x2
It seems that the placement of the accelerometer on our
board does not belong to any default, so fih added the 9th
place, counting from 0 to 8 it is the latter one.

{  2,    1,    0,     1,      1,     -1 }, /* P8 */

I also checked the place sysfs attribute when the driver was
working before, and I got 8.
In a recent commit, we started using the parse dt function from
caf instead of the fih one, which now parses the bosch,place
property, and as we hadn't specified any, it was setting 0 as
default, so we were getting something like this:

Invalid place parameter, use default value 0

Which is wrong because we use the nineth one, and the rotation
would work weird.

We also need to modify the driver in order to this non-specific
value to be taken into account.

Reference:
https://github.com/danielgusvt/destiny/blob/staging141/Documentation/devicetree/bindings/input/misc/bma2x2.txt#L16-L17