arm: DT: MSM8939: Fix/update bma2x2 node for caf driver
In commit "input: sensors: add board specific configration for
BOSCH bma2x2", we started using the parse dt function from caf
instead of the fih one, which now parses the bosch,place
property, instead of bosch,layout so we now have to replace it
in order to reflect the driver change, otherwise it will set 0
as default and the driver would not work correctly.

Also change the pinctrl nodes and names to follow the
other bma2x2 devices, because we now use the caf's pinctrl
stuff; disable the interrupt stuff because it is optional and
fih seemed to forcefully disable it.

Finally, remove the bosch,vdd-current-level property, because
we never used it.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>