arm: DT: msm8939-mtp: Fix/update bma2x2 node for caf driver
The placement of the accelerometer on our board does not belong
to any default, so fih added the 9th place, counting from 0 to
8 it is the latter one.

{  2,    1,    0,     1,      1,     -1 }, /* P8 */

In commit "input: sensors: add board specific configration for
BOSCH bma2x2", we started using the parse dt function from caf
instead of the fih one, which now parses the bosch,place
property, instead of bosch,layout so we now have to replace it
in order to reflect the driver change, otherwise it will set 0
as default and the driver would not work correctly.

We of course need to modify the driver as well, in order to use
this non-general value.

Finally, remove the bosch,vdd-current-level property, because
we don't use it, update the pinctrl-names and use gpio-int1
instead of int_pin because we now use the caf's pinctrl stuff.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>