[TEMP] input: bma2x2: Change driver name
From now on, we will be using the hal+sensord from AOSP because
it eases my development. How? I had to hack this driver for
this to work, because the sensor binaries are ancient and
don't want to get along with the new code. Moreover,
having hacks + and adding new code is not a good idea. Also,
my purpose for this driver is to work on AOSP roms.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>