leds: Add sh1tty lp5562 fxn implementation
Why am I doing this? Because the stock implementation that FIH has
followed is really crappy and unknown for me, and therefore I
couldn't get the regular driver to work, which does in AOSP.
Maybe one day I can get it to work, but it is not my priority now.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>