input: touchscreen: cyttsp4: Complete RestoreGloveState workaround
This workaround is used to restore the state of the glove mode.
The AOSP driver had part of this workaround in an incomplete way.
Without these hooks, the glove mode was always enabled by default
and stubborn if we disable it via sysfs or the settings app in the
stock rom.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>