input: ts: cyttsp4/5: Kconfig: Declare I2C as dependency instead of select
If the I2C config is selected instead of being declared as what
it actually is (a dependency), the build system in the kernel will
not take it into account for the defconfig (and will remove it
in a make savedefconfig), which works fine if we keep
these touchscreen drivers enabled.

But, if the whole touchscreen drivers are disabled, the I2C config
won't be brought back, and the compilation will fail because we have
other drivers that rely on this subsystem and declare this dependency
as 'depends on', which is the correct way.

See:
https://github.com/sonyxperiadev/kernel/commit/281c3b7bd905e451927dfb6c7766b2cbe271e015#diff-8595c3af646bd1fa38b0820b61bf40d7
https://github.com/sonyxperiadev/kernel/commit/1342e27196dc161b3da260cbb1dd011d2bfd8fce#diff-8595c3af646bd1fa38b0820b61bf40d7

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>